VERSION 2.0.4
Battery charge problem is tried to be solved. When battery is totally discharged and USB is plugged in, processor goes to boot mode and battery cannot be charged. In order to solve this problem, PCB ECG_Du_22 is manually reconnected. Boot pin is connected directly to ground, and rest part of boot lead is connected to BT_UART_TX_IND pin (previously disconnected from BLE) as USB detection.